fix(mangadex): move MangaDexClient to types.ts, drop dead PageReport, cache at-home server, fix missing await
- MangaDexClient interface moved from index.ts to types.ts; re-exported via export type
- Deleted PageReport (duplicate of ReportPayload, never imported)
- mangadexImageFetcher now fetches at-home server once before the loop; re-fetches only on retry failure path
- Added missing await on rejects.toThrow assertion in client.test.ts
